chore: ToString国际化与全局using优化，提升健壮性
- 统一ToString方法为CultureInfo.InvariantCulture，保证格式一致性
- 常用命名空间通过GlobalUsing全局引入，减少重复
- 清理多余using，代码更简洁
- 优化资源释放逻辑，防止泄漏，提升异步释放正确性
- 日志格式国际化
- 屏蔽未用私有字段警告
- 细节逻辑优化（如订阅信用额度回绕）
整体提升代码国际化兼容性、资源管理健壮性和可维护性

+++ src/EasilyNET.Drivers.S7Plus/S7CommPlus/Subscriptions/Subscription.cs
@@ -1,6 +1,5 @@
 // SPDX-License-Identifier: LGPL-3.0-or-later
 // Derived from thomas-v2/S7CommPlusDriver, Copyright (C) 2023 Thomas Wiens. See LICENSE-LGPL-3.0.txt.
-using Microsoft.Extensions.Logging;
 using EasilyNET.Drivers.S7Plus.S7CommPlus.ClientApi;
 using EasilyNET.Drivers.S7Plus.S7CommPlus.Core;
 using EasilyNET.Drivers.S7Plus.S7CommPlus.Net;
@@ -245,7 +244,7 @@ internal sealed partial class S7CommPlusConnection
                     // 信用额度为 0 等同“不再发送任何通知”，回绕到 0 会让长跑订阅卡死：跳过 0 保持数据流
                     if (m_NextCreditLimit <= 0)
                     {
-                        m_NextCreditLimit = (short)creditLimitStep;
+                        m_NextCreditLimit = creditLimitStep;
                     }
                     if (log.IsEnabled(LogLevel.Debug))
                     {
